requirements.txt
alembic==1.0.0
APScheduler==3.5.3
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
blinker==1.4
boto3==1.9.5
botocore==1.12.5
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
configparser==3.5.0
cookies==2.2.1
coverage==4.5.1
docutils==0.14
Flask==1.0.2
Flask-APScheduler==1.10.1
Flask-GoogleMaps==0.2.6
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
geographiclib==1.49
geopy==1.17.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Logbook==1.4.0
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
PyJWT==1.6.4
PySocks==1.6.8
pytest==3.8.0
pytest-cov==2.6.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
responses==0.9.0
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.12
twilio==6.19.1
tzlocal==1.5.1
undecorated==0.3.0
urllib3==1.23
Werkzeug==0.14.1
facebook-sdk==3.0.0